transaction bug fix for returning an error, renaming Price to PricingType
